Fixed the test for Statistics::Write to conform with the change of making numTuples into doubles

+++ include/Statistics.h
@@ -14,7 +14,7 @@
 
 struct StatPair {
 	double numTuples;
-	std::unordered_map<std::string, double> atts;
+	std::unordered_map<std::string, int> atts;
 };
 
 class Statistics {
